% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odin_dust.R
\name{odin_dust}
\alias{odin_dust}
\alias{odin_dust_}
\title{Create a dust odin model}
\usage{
odin_dust(x, ..., options = NULL)

odin_dust_(x, ..., options = NULL)
}
\arguments{
\item{x}{Either the name of a file to read, a text string (if
length is greater than 1 elements will be joined with newlines)
or an expression.}

\item{...}{Arguments passed to \link{odin_dust_options},
including \code{real_type}, \code{gpu}, \code{verbose}, \code{workdir},
\code{no_check_unused_equations} and \code{rewrite_dims}.}

\item{options}{An \link[odin:odin_options]{odin::odin_options} or \link{odin_dust_options}
object. If given it overrides arguments; if it is already a
\code{odin_dust_options} object it is returned unmodified. Otherwise
it is passed through to \link[odin:odin_options]{odin::odin_options} where it will
override arguments in \code{...} but respond to the \code{odin_dust}
specific options (\code{real_type}, etc)}
}
\description{
Compile an odin model to work with dust.
}
\details{
Note that this does not (yet) support the full odin output object,
instead creating the more limited dust interface. However, for
many uses this should be considerably faster than the interface
that odin normally uses (built on dde).
}
